+++ src/main/java/org/philosophy/carwashing/dto/requestdto/WashTypeRequestDto.java
@@ -2,7 +2,6 @@ package org.philosophy.carwashing.dto.requestdto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.Duration;
 
 @Getter
@@ -14,6 +13,6 @@ public class WashTypeRequestDto {
 
     private String name;
     private Duration duration;
-    private BigDecimal cost;
+    private Double cost;
 
 }
