+++ src/main/java/org/philosophy/carwashing/dto/responsedto/BookingResponseDto.java
@@ -3,7 +3,6 @@ package org.philosophy.carwashing.dto.responsedto;
 import lombok.*;
 import org.philosophy.carwashing.enums.BookingStatuses;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +14,7 @@ public class BookingResponseDto {
 
     private Integer id;
     private Integer requestId;
-    private BigDecimal totalCost;
+    private Double totalCost;
     private BookingStatuses status;
     private Boolean isPaid;
     private LocalDateTime paymentTime;
