+++ src/main/java/org/philosophy/carwashing/dto/responsedto/WashTypeResponseDto.java
@@ -2,7 +2,6 @@ package org.philosophy.carwashing.dto.responsedto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.Duration;
 
 @Getter
@@ -15,6 +14,6 @@ public class WashTypeResponseDto {
     private Integer id;
     private String name;
     private Duration duration;
-    private BigDecimal cost;
+    private Double cost;
 
 }
